When saerching terms in the DPV vocabularies, try no just the dpv namespace, but also dpv-pd and dpv-gdpr.
Updated the error message for a missing ‘action’ parameter in the
query.

When a property of a DPV term has an empty list of values, do not
output the property at all, rather than output it with an empty list
as value.

Added a query for known controllers (organizations): given the ID of a
controller, return information about that controller.

Added a query for the logo of a controller: given the ID of a
controller, return an image. (unlike the other queries, the return is
not a bit of JSON-LD, but an image.)

Added the controller and logo queries to the debug form.